% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{AAtranslate}
\alias{AAtranslate}
\title{Translate Codons into Amino Acids}
\usage{
AAtranslate(x)
}
\arguments{
\item{x}{A character string containing ambiguity codes for a codon.}
}
\value{
A list containing:
\item{Codons}{A list of possible codons resulting from the ambiguous codon}
\item{AAs}{A character string of the possible amino acids that could result,
  written as single letter AA codes}
}
\description{
A function to calculate possible Amino Acid codes
from a codon containing  ambiguity codes.
}
\details{
Translating a diploid DNA sequence into an AA sequence.
}
\examples{
AAtranslate("WYR")
}
\references{
Bill Black IV
}
\seealso{
[DNAtranslate()], [ambiguityCodons()], [codons()]
}
\author{
Stu Field and Bill Black IV
}
