% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ago2}
\alias{Ago2}
\title{SNP data}
\format{
Data frame SNP data used as an example for [findSNPs()].
}
\source{
Stu Field
}
\description{
SNP data for the Ago2 gene. Used in LinkDig exercise, aka
Population Genetics class at Colorado State Univ. with Bill Black.
}
\references{
Stu Field, Bill Black IV
}
