% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{allSame}
\alias{allSame}
\title{Are All Entries Identical?}
\usage{
allSame(x)
}
\arguments{
\item{x}{A numeric or character vector or a matrix that could be converted
to a vector.}
}
\value{
A logical value indicating if all values in the vector are identical.
}
\description{
Determine if all entries of a vector are identical. Can work on a matrix
also because matrix could be converted to a vector.
}
\examples{
allSame(1:15)
allSame(rep("DOG", 10))
allSame(diag(5))
}
\author{
Stu Field
}
