% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambiguityCodons}
\alias{ambiguityCodons}
\title{Ambiguity codons}
\format{
Data frame containing all the ambiguity codons.
}
\source{
Stu Field
}
\description{
Codon chart for all the single letter codon representations.
}
\references{
Stu Field, Bill Black
}
