% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{array2df}
\alias{array2df}
\title{Array Data to Data Frame}
\usage{
array2df(x, names, Meta = c(1, 1), probechar = 13)
}
\arguments{
\item{x}{The "RGList" class object that is the result of
a call to [limma::RG.MA()].}

\item{names}{The column names of `x` that you wish to
include in the order you want them.}

\item{Meta}{The row (1) and col (2) of MetaRow and MetaCol
as seen in original R-code.
Not always used so default is set to c(1, 1).}

\item{probechar}{The cut off for the length of the probe name (default = 13)}
}
\value{
A data frame containing all the desired array data.
}
\description{
Reorganize array data into a data frame. Uses \code{\link{cbind}} to combine
arrays horizontally with Probe ID running down column 1. Each array
red-signal green-signal weighting for probe running horizontally for each of
the arrays
}
\details{
Bind all the arrays together horizontally i.e. probe id running down
column 1 and each array red signal-green signal-weighting for probe
running horizontally for each of the arrays

Takes as input x, the RG.corrected limma class object
Meta = 1) MetaRow, 2) MetaCol as included in original R-code
}
\examples{
RG.corrected$targets
colNames <- c("ProbeName", "GeneName", "SystematicName",
              "Description", "Status", "Sequence", "Row", "Col")
Tororo.raw <- array2df(RG.corrected, names = colNames)
head(Tororo.raw, 25)
}
\seealso{
[limma()]
}
\author{
Stu Field
}
