% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{boot.pop}
\alias{boot.pop}
\title{Bootstrap Phenotypes}
\usage{
boot.pop(x, loci, pops, phenotypes, rseed = sample(1000, 1))
}
\arguments{
\item{x}{The original vector of population data}

\item{loci}{The loci present in the population}

\item{pops}{A factor variable of the populations}

\item{phenotypes}{Phenotypes of the population. Usually "R" and "S".}

\item{rseed}{Optional for reproducibility}
}
\value{
A bootstrap population based on the original population data
}
\description{
Create bootstrap population of phenotypes
}
\examples{
# add example
\dontrun{a <- 1}
}
\author{
Stu Field and Martin Donnelly
}
