% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{DNAtranslate}
\alias{DNAtranslate}
\title{DNA Translate}
\usage{
DNAtranslate(x, csv = FALSE, file)
}
\arguments{
\item{x}{A vector of DNA sequence for translation into AA sequence.}

\item{csv}{Logical. Create a csv output file?}

\item{file}{Character. Name of the csv output file if desired.}
}
\value{
A list containing:
\item{DNAsequence}{Original DNA sequence.}
\item{AAsequence}{Translated AA sequence based on DNA sequence.}
\item{TranslationTable}{Table combining in nice format.}
}
\description{
Translate a sequence of DNA nucleotides to AA sequence.
Ambiguity codes are allowed.
}
\examples{
chunk1 <- c("YYYAAGATCGTCGGTGGCGATGAGGCCGAAGCGCACGAATTTCCCTACCAAATCTCG")
chunk2 <- c("CTGCAGTGGAACTTCAACGATGGACAAACGGAGACCATGCACTTCTGYGGAGCTTCGG")
chunk3 <- c("TGTTGAACGAAAACTTYGTCCTGACGGCTGCTCACTGCAAGACCGCATACTCCAATA")
chunk4 <- c("CCGGGTWCATCGAAGTGGTTGCCGCTGAACATGATGTGGCYGTTGCGGAAGGATCCGA")
chunk5 <- c("ACAGCGTCGYTTGGTTGCGGAGTTCATCGTCCACGAGGACTATCAAGGRGGAGTCAGT")
chunk6 <- c("CCCGATGAGATTGCCGTCA")
DNAseq <- paste0(chunk1, chunk2, chunk3, chunk4, chunk5, chunk6)
DNAseq

AAtranslate("YRY")
DNAtranslate(DNAseq, csv = FALSE)
}
\seealso{
[AAtranslate()], [searchReplace()]
}
\author{
Stu Field, Martin Donnelly and Bill Black
}
