% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{findSNPs}
\alias{findSNPs}
\title{Find and Calculate SNPs}
\usage{
findSNPs(x)
}
\arguments{
\item{x}{A matrix of raw SNP data, individuals as rows & loci as cols.}
}
\value{
A list containing:
\item{Data}{The raw data reorganized as a matrix. Individuals as rows.
  Loci as columns.}
\item{Individuals }{Number of individuals in the data set. Equal to `nrows`.}
\item{Length}{Length of the SNP sequence. Equal to ncol}
\item{nSNP}{Number of SNPs in the data set}
\item{SNPs}{Position of the SNP loci, identified by the col number}
}
\description{
Determine the relative positions of SNPs for variable SNP data.
Individuals are rows and nucleotide positions are cols.
Must be a data frame or a matrix. See Examples.
}
\details{
From Diploid genetic sequences.
}
\examples{
Ago2
findSNPs(Ago2)
}
\seealso{
[allSame()]
}
\author{
Stu Field, Bill Black
}
\keyword{SNP}
