% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{formatMicrosat}
\alias{formatMicrosat}
\title{Convet Microsatellite Data Format}
\usage{
formatMicrosat(rawdata, col.loci, locus.names)
}
\arguments{
\item{rawdata}{Raw microsattellite data containing separate columns for each
of the diploid loci. Converts to a "/" separating the diploid alleles.}

\item{col.loci}{Defining the columns which contain microsat data.}

\item{locus.names}{Names of the microsat loci}
}
\value{
Matrix of microsatellite data with "/" as the allele separator,
rather than separate cols for alleles.
}
\description{
Convert adjacent microsatellite data from separate columns for each of the
diploid loci to a "/" separating the alleles (i.e. allele1/allele2).
}
\examples{
head(HarrCraigData)
Locusnames <- c("L3RIH59", "L3RIND30", "L3RIR01", "L3RIR02",
                "L3RIR03", "L3RIR04", "L3RIR05", "L3RIR07",
                "L3RIR08", "L3RIR09", "L3RIR10", "L3RIR12")

# Reformat
FormattedData <- formatMicrosat(HarrCraigData, col.loci = 4:27,
                                locus.names = Locusnames)
head(FormattedData)

}
\references{
"inst/HarrCraig.csv"
}
\seealso{
[HarrCraigData]
}
\author{
Stu Field
}
