% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ghana_MvS_SNPdata}
\alias{Ghana_MvS_SNPdata}
\title{Ghana M vs. S SNP Data}
\format{
A data frame with 871 cases and the following headings: \cr
"Chrom" \cr "Position" \cr "ScaledPosition" \cr
"Case_F_Allele1"\cr"Case_F_Allele2"\cr"Control_F_Allele1"\cr"Control_F_Allele2"\cr
"CaseFreq"\cr"ControlFreq"\cr"Chisq"\cr"P"\cr"negLogP"\cr"Fis"\cr"Fst"
}
\source{
Stu Field
}
\description{
Used in the sliding window analysis.
}
\references{
Stu Field, Martin Donnelly
}
\keyword{SNP}
\keyword{data.}
