% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GhanaKDR}
\alias{GhanaKDR}
\title{GhanaKDR Data}
\format{
Data frame in the construction of a lineage map.
}
\source{
Stu Field
}
\description{
GhanaKDR of SNP data for 60 individuals and SNPs as cols.
Core was set to 25 (i.e. the 25th nucleotide).
}
\references{
Stu Field, Martin Donnelly
}
