% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HarrCraigData}
\alias{HarrCraigData}
\title{Harr Craig Resistance Data}
\format{
Data frame used in the [formatMicrosat()] function to combine
cols to allele1/allele2 format.
}
\description{
Microsatellite data for R/S phenotype data in mosquitos.
}
\references{
Stu Field, Martin Donnelly, Harr Craig
}
