% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{HeBootstrapFUN}
\alias{HeBootstrapFUN}
\title{Heterozygosity Bootstrapping}
\usage{
HeBootstrapFUN(x, nboot, LociCols)
}
\arguments{
\item{x}{Microsatellite data object (usually data frame). See example.}

\item{nboot}{Number of bootstrap replicates.}

\item{LociCols}{Which cols are microsatellite loci.}
}
\value{
A list containing:
\item{estimate }{The empirical estimate of heterozygosity by population
and by phenotype}
\item{upperCI9 }{upper confidence limit of the heterozygosity estimate}
\item{lowerCI95 }{lower confidence limit of the heterozygosity estimate}
}
\description{
Conduct bootstrap analysis on heterozygosity microsatellite data for
resistance/susceptible phenotypes. Originally written for the analysis of
Harr Craig data set.
}
\details{
Bootstrapping routine on original MicroData: HarrCraigData
}
\note{
Bootstrapping routine on original MicroData
}
\examples{
Locusnames <- c("L3RIH59", "L3RIND30", "L3RIR01", "L3RIR02",
                "L3RIR03", "L3RIR04", "L3RIR05", "L3RIR07",
                "L3RIR08", "L3RIR09", "L3RIR10", "L3RIR12")

MicroData <- formatMicrosat(HarrCraigData, col.loci = 4:27, locus.names = Locusnames)

# exclude Alajo and Malawi because:
# no Alajo|S and no Malawi|R
# cannot estimate heterozygosity with 1 phenotype
MicroData <- subset(MicroData, Population != "Alajo" & Population != "Malawi")
head(MicroData)

HeBootstrapFUN(MicroData, nboot = 100, LociCols = 4:15)

}
\references{
Harr Craig microsatellite data
}
\seealso{
\code{\link{HeFUN}}, \code{\link{boot.pop}}
}
\author{
Stu Field, Martin Donnelly
}
