% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KenyaSNP_EHHdata}
\alias{KenyaSNP_EHHdata}
\title{Lineage Map of Kenya SNP data}
\format{
A data frame
}
\description{
Lineage map used in the EHH analysis.
}
\references{
Stu Field, Martin Donnelly
}
