% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{lineageMap}
\alias{lineageMap}
\title{Lineage Map & EHH Analysis}
\usage{
lineageMap(data, core, csv = FALSE, file)
}
\arguments{
\item{data}{Raw data of nucleotide changes.}

\item{core}{Set the nucleotide position of the core.}

\item{csv}{Write a csv output file of the resulting lineage map?}

\item{file}{Character. If `csv = TRUE`, the name of the output file.
Otherwise ignored.}
}
\value{
A list containing:
\item{SNP.Data}{Original SNP data.}
\item{Haplotypes}{The number of haplotypes lineages left & right of the
core}
\item{Lineages}{Table showing the number of haplotypes in each
lineage}
\item{Haplo.Lineages}{The actual lineage map; ordered from top
to bottom}
}
\description{
Creates a lineage map of a given data set as part of the Extended Haplotype
Heterozygosity (EHH) analysis.
}
\examples{
names <- paste(GhanaKDR[, 1], GhanaKDR[, 2], sep = "")
new.data <- GhanaKDR[-c(1, 2)]
colnames(new.data) <- 1:ncol(new.data); rownames(new.data) <- names
new.data <- new.data[-59,]    # remove resistant individual
head(new.data)

# replace numerics with nucleotides
new.data <- searchReplace(new.data, s = 1:4 , r = c("A","C","G","T"))

# the call
lineageMap(new.data, core = 25)

\dontrun{
lineageMap(new.data, core = 25, csv =TRUE, file = "lineageOutput.csv")
}
}
\seealso{
[allSame()], [SNPboot()]
}
\author{
Stu Field, Martin Donnelly
}
