% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microsatdata}
\alias{microsatdata}
\title{Microsatellite Data}
\format{
Data frame microsatellite data using the "/" format.
}
\description{
Microsatellite data as given in the pegas example 
for 237 individuals and 9 loci.
}
\references{
Stu Field, Emmanuel Paradis
}
\keyword{Pegas}
