% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%<>\%}
\title{Pipe operators}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} and
\code{magrittr::\link[magrittr:pipe]{\%<>\%}} for details.
}
\keyword{internal}
