% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{plotSlideSNP}
\alias{plotSlideSNP}
\title{Sliding SNP Plot}
\usage{
plotSlideSNP(
  x,
  y,
  breaks,
  ylabel,
  font,
  line.cols,
  line.lty,
  save = FALSE,
  ...
)
}
\arguments{
\item{x}{Midpoints. Typically the output of `slide.window.bp()$midpoints`.}

\item{y}{Values. Typically the output of `slide.window.bp()$values`.}

\item{breaks}{Base pairs where the chromosome breaks occur from one
chromosome to the next.}

\item{ylabel}{Label for the y-axis.}

\item{font}{Font for the plot, passed to [par()].}

\item{line.cols}{Line colours for plotting.}

\item{line.lty}{Line type for plotting}

\item{save}{Logical.}

\item{...}{Additional arguments passed to [plot()].}
}
\value{
A step "sliding window" plot.
}
\description{
General wrapper for plotting sliding window analyses on SNP
chromosome base pair positions. Primary utility in [slide.window.bp()].
}
\note{
Length of `breaks =`, `line.cols =`, and `line.lty =`
*must* all be equal.
}
\examples{
head(Ghana_MvS_SNPdata)

BPwin    <- 7500000
negLogP  <- Ghana_MvS_SNPdata$negLogP
data.out <- slide.window.bp(negLogP, Pos = Ghana_MvS_SNPdata$ScaledPosition,
                            bp = BPwin, FUN = max)
data.out

chr.breaks <- c(60707344, 62812924, 110916950, 162910923, 206568235)
chr.cols   <- c(2, 2, 3, 2, 3)
chr.lines  <- c(2, 2, 1, 2, 1)

plotSlideSNP(x = data.out$midpoints, y = data.out$values, type = "s",
             col = "navy",
             lwd = 1.25,
             ylabel = "-logP (max)",
             breaks = chr.breaks,
             font = 4,
             line.cols = chr.cols,
             line.lty = chr.lines,
             save = FALSE)
}
\seealso{
[slide.window.bp()]
}
\author{
Stu Field, Martin Donnelly
}
