% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{slide.window.bp}
\alias{slide.window.bp}
\title{Sliding Window SNP Analysis (bp)}
\usage{
slide.window.bp(X, Pos, bp, FUN)
}
\arguments{
\item{X}{Vector object containing data appropriate to be applied by
`FUN`.}

\item{Pos}{Vector of base pair (SNP) positions to evaluate.}

\item{bp}{Size of the window over which the `FUN` should be evaluated
in number of base pairs. Equivalent to `window =` in
[slide.window()].}

\item{FUN}{The summary function that evaluates the numbers in the window,
typically `mean` or `max`.}
}
\value{
A list containing:
\item{values}{The result of the application of `FUN` in each of the windows.}
\item{midpoints}{The midpoint of the window, used primarily for plotting.}
\item{n.SNPs}{How many SNPs evaluated in each window based on constant bp
jumps along the chromosome.}
}
\description{
Sliding window analysis for SNPs along chromosome based on base pair
positions rather than SNP positions. Could be any vector of values.
}
\examples{
r.vec <- sample(1:20, 100, replace = TRUE)  # random vector
slide.window.bp(X = r.vec, Pos = 1:length(r.vec), bp = 5, FUN = mean)

head(Ghana_MvS_SNPdata)

BPwin   <- 7500000
negLogP <- Ghana_MvS_SNPdata$negLogP
plot.data <- slide.window.bp(negLogP, Pos = Ghana_MvS_SNPdata$ScaledPosition,
                             bp = BPwin, FUN = max)

chr.breaks <- c(60707344, 62812924, 110916950, 162910923, 206568235)
chr.cols <- c(2, 2, 3, 2, 3)
chr.lines <- c(2, 2, 1, 2, 1)

plotSlideSNP(x = plot.data$midpoints, y = plot.data$values,
             type = "s", col = "navy",
             lwd = 1.25, ylabel = "-logP (max)",
             breaks = chr.breaks, font = 4,
             line.cols = chr.cols,
             line.lty = chr.lines,
             save = FALSE)
}
\seealso{
[plotSlideSNP()], [slide.window()]
}
\author{
Stu Field, Martin Donnelly
}
