% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{slide.window}
\alias{slide.window}
\title{Sliding Window SNP Analysis}
\usage{
slide.window(X, window, frame.skip = 1, FUN)
}
\arguments{
\item{X}{Vector object containing data appropriate to be applied by
\code{FUN}.}

\item{window}{Size of the window over which the \code{FUN} should be
evaluated.}

\item{frame.skip}{How far the window should jump from one iteration to the
next. Default = 1.}

\item{FUN}{The summary function that evaluates the numbers in the window,
typically \code{mean} or \code{max}.}
}
\value{
A list containing:
\item{values}{The result of the application of \code{FUN} in each of the windows.}
\item{midpoints}{The midpoint of the window, used primarily for plotting.}
\item{length}{How many 'windows' were evaluated during the analysis.}
}
\description{
Sliding window analysis for SNPs along chromosome.
Can be any vector of values.
}
\examples{
# Example 1
r.vec <- sample(1:20, 100, replace = TRUE)       # create random vector
slide.window(X = r.vec, window = 5, frame.skip = 2, FUN = max)

# Example 2
head(Ghana_MvS_SNPdata)

Win <- 5
negLogP <- Ghana_MvS_SNPdata$negLogP
data.out <- slide.window(negLogP, window = Win, frame.skip = 2, FUN = max)

plot(data.out$values ~ data.out$midpoints,
     type = "s", col = "navy",
     ylab = "-logP (max)",
     xlab = "midpoints", lwd = 1.5)
legend("topleft",
       legend = format(paste("WindowSize =", Win)),
       bg = "gray75", cex = 0.75, box.lty = 0)
}
\seealso{
[lotSlideSNP()], [slide.window.bp()]
}
\author{
Stu Field, Martin Donnelly
}
