% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito.R
\name{SNPboot}
\alias{SNPboot}
\title{SNP Bootstrap of EHH data}
\usage{
SNPboot(
  x,
  nBoot,
  core,
  lo = 0.025,
  up = 0.975,
  pts = 19,
  hist = FALSE,
  plot = TRUE,
  csv = FALSE,
  file
)
}
\arguments{
\item{x}{A data frame or matrix with haplotype lineage data, typically
output from [lineageMap()]. Column 1 MUST be a character vector,
usually with an Amino Acid string describing the locus.}

\item{nBoot}{Number of bootstrap samples to perform for each SNP locus.}

\item{core}{Set which column (SNP) should be considered the `core`.
Typically determined from [lineageMap()].}

\item{lo}{Lower confidence limit passed to [quantile()].}

\item{up}{Upper confidence limit passed to [quantile()].}

\item{pts}{Size of the points for the EHH plot.}

\item{hist}{Logical. Should a histogram also be plotted showing the
distribution of the bootstraps of EHH?}

\item{plot}{Logical. Should the Expected Haplotype Homozygosity be plotted?}

\item{csv}{Logical. Should the data frame containing EHH point estimate &
CI95 bootstrap errors be saved to a file?}

\item{file}{Character. If `csv = TRUE`, the name of the *.csv file to be
produced. Otherwise ignored.}
}
\value{
A data frame (with `nrow = nSNP`), consisting of:
\item{Est H-H}{the EHH point estimate}
\item{2.5\%}{the lower CI95}
\item{97.5\%}{the upper CI95}

This data frame is then used by the [gplots::plotCI()] function
plot the EHH plot with bootstrap CI95 errors.
}
\description{
Perform EHH analysis on SNP data and calculate CI95
via bootstrapping method.
}
\details{
1st col *must* be character class!
}
\note{
Requires the \pkg{gplots} package.

If the bootstrap confidence interval is very short
relative to the point estimate, [gplots::plotCI()] will give the
warning: `In arrows(...): zero-length arrow is of indeterminate
angle and so skipped`.
}
\examples{
head(KenyaSNP_EHHdata)   # pkg data

# Subset data sets by Leu/Ser
data.Leu <- subset(KenyaSNP_EHHdata, KenyaSNP_EHHdata == "Leu")
data.Ser <- subset(KenyaSNP_EHHdata, KenyaSNP_EHHdata == "Ser")

SNPboot(data.Leu, nBoot = 50, core = 21)  # nBoot >500 recommended
SNPboot(data.Ser, nBoot = 25, core = 21)
}
\references{
"Kenyan_SNP_EHH_Data.csv"
}
\seealso{
[boot.pop()], [HeBootstrapFUN()], [gplots::plotCI()],
[lineageMap()], [quantile()]
}
\author{
Stu Field, Martin Donnelly
}
