% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tororo}
\alias{Tororo}
\alias{RG.corrected}
\title{Tororo Microarray Data}
\format{
Limma object.
}
\description{
Microarray data from Tororo mosquitos, R/S phenotype, used in the
\pkg{limma} package.
}
\references{
Stu Field, Martin Donnelly
}
